Console: Update rrd4j to version 3.9
(cherry picked from commit b0db90da49854ca060ea7d693adc4bfe4e587612)

+++ apps/jrobin/java/src/org/rrd4j/graph/Mapper.java
@@ -25,7 +25,7 @@ class Mapper {
             pixieY = im.ysize / (im.maxval - im.minval);
         }
         else {
-            pixieY = (double) im.ysize / (Math.log10(im.maxval) - Math.log10(im.minval));
+            pixieY = im.ysize / (Math.log10(im.maxval) - Math.log10(im.minval));
         }
     }
